The great API-creation refactoring
This is the first in probably a long sequence of refactorings that add the
backend agnostic API. While a buch of new stuff is added and a significant
amount of changes, none of them are actually in use at the moment.

The current state is commited as is, because at this point it can compile
and run, and I don't think I can make more changes without breaking stuff.